move auth to middleware
