+++ storage/ci/nginx.conf
@@ -34,20 +34,22 @@ http {
         port_in_redirect off;
         server_name localhost;
 
+        add_header 'Access-Control-Allow-Origin' '*';
+        add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS, PUT, DELETE';
+        
         location / {
-            if ($request_method = 'OPTIONS') {
-                add_header 'Access-Control-Allow-Origin' '*';
-                add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS, PUT, DELETE';
+                if ($request_method = 'OPTIONS') {
+                   add_header 'Access-Control-Allow-Origin' '*';
+                   add_header 'Access-Control-Allow-Methods' 'GET, POST, OPTIONS, PUT, DELETE';
+                   # Custom headers and headers various browsers *should* be OK with but aren't
+                   add_header 'Access-Control-Allow-Headers' 'DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Content-Range,Range,publicKey,Authorization';
 
-                # Custom headers and headers various browsers *should* be OK with but aren't
-                add_header 'Access-Control-Allow-Headers' 'DNT,X-CustomHeader,Keep-Alive,User-Agent,X-Requested-With,If-Modified-Since,Cache-Control,Content-Type,Content-Range,Range,Authorization';
-
-                # Tell client that this pre-flight info is valid for 20 days
-                add_header 'Access-Control-Max-Age' 1728000;
-                add_header 'Content-Type' 'text/plain charset=UTF-8';
-                add_header 'Content-Length' 0;
-                return 204;
-            }
+                   # Tell client that this pre-flight info is valid for 20 days
+                   add_header 'Access-Control-Max-Age' 1728000;
+                   add_header 'Content-Type' 'text/plain charset=UTF-8';
+                   add_header 'Content-Length' 0;
+                   return 204;
+                }
 
             root   /app/api/public;
             index  index.php index.html index.php;
